# Conflicts:
#	config/packages/workflow.yaml
#	config/services.yaml
#	src/Classes/Export/ExportBcc.php
#	src/Classes/ValidationProcess.php
#	src/Command/MigrateTroncCommunSemestresCommand.php
#	src/Controller/AdminEditController.php
#	src/Controller/FicheMatiereController.php
#	src/Security/Voter/RessourceVoter.php
#	src/Service/Semestre/TroncCommunToMutualisationMigrator.php
#	templates/components/CommentaireAdmin.html.twig
#	templates/typeDiplome/formation/_admission.html.twig
#	templates/typeDiplome/formation/_contacts.html.twig
#	templates/typeDiplome/formation/_descriptifParcours.html.twig
#	templates/typeDiplome/formation/_etApres.html.twig
#	templates/typeDiplome/formation/_inscription.html.twig
#	templates/typeDiplome/formation/_localisationParcours.html.twig
#	templates/typeDiplome/formation/_presentationFormation.html.twig
#	translations/process.fr.yaml

+++ src/Controller/NotificationController.php
@@ -3,9 +3,12 @@
 namespace App\Controller;
 
 use App\Classes\JsonReponse;
+use App\Entity\Notification;
+use App\Entity\User;
 use App\Repository\NotificationRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
@@ -18,7 +21,7 @@ class NotificationController extends AbstractController
         Request                $request
     ): Response {
         $notification = $notificationRepository->find($request->query->get('id'));
-        if ($notification !== null) {
+        if ($notification !== null && $this->canAccess($notification)) {
             $notification->setLu(true);
             $notificationRepository->save($notification, true);
 
@@ -50,6 +53,9 @@ class NotificationController extends AbstractController
     ): Response {
         $notifications = $notificationRepository->findBy(['destinataire' => $this->getUser()], ['created' => 'DESC']);
         foreach ($notifications as $notification) {
+            if ($notification->requiresAck()) {
+                continue;
+            }
             $notificationRepository->remove($notification);
         }
         $entityManager->flush();
@@ -67,5 +73,37 @@ class NotificationController extends AbstractController
         return $this->render('notification/_liste.html.twig', [
             'notifications' => $notifications,
         ]);
+    }
+
+    #[Route('/notification/mutualisation/{id}/valider', name: 'app_notification_mutualisation_valider', methods: ['GET'])]
+    public function validerMutualisation(
+        Notification           $notification,
+        NotificationRepository $notificationRepository,
+        Request                $request,
+    ): RedirectResponse
+    {
+        if (!$this->canAccess($notification)) {
+            throw $this->createAccessDeniedException();
+        }
+
+        $notification->setLu(true);
+        $notificationRepository->save($notification, true);
+
+        $referer = $request->headers->get('referer');
+        if ($referer !== null && $referer !== '') {
+            return $this->redirect($referer);
+        }
+
+        return $this->redirectToRoute('app_homepage');
+    }
+
+    private function canAccess(Notification $notification): bool
+    {
+        $user = $this->getUser();
+        if (!$user instanceof User) {
+            return false;
+        }
+
+        return $notification->getDestinataire()?->getId() === $user->getId();
     }
 }
